requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
ytmusicapi>=1.0.0
python-dotenv>=1.0.0
drf-yasg>=1.21.0

